Moving Models to a new project
Init AlgorithmConverter
