--- JiraSchedulingConnectAppService/Models/AccountRole.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace JiraSchedulingConnectAppService.Models;
-
-public partial class AccountRole
-{
-    public string? AccountId { get; set; }
-
-    public int Id { get; set; }
-
-    public int? TokenId { get; set; }
-
-    public bool? IsDelete { get; set; }
-
-    public DateTime? CreateDatetime { get; set; }
-
-    public bool? IsDelete1 { get; set; }
-
-    public DateTime? DeleteDatetime { get; set; }
-
-    public virtual AtlassianToken? Token { get; set; }
-}
