+++ ModelLibrary/DBModels/TaskPrecedence.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace ModelLibrary.DBModels
+{
+    public partial class TaskPrecedence
+    {
+        public int TaskId { get; set; }
+        public int PrecedenceId { get; set; }
+        public DateTime? CreateDatetime { get; set; }
+        public bool? IsDelete { get; set; }
+        public DateTime? DeleteDatetime { get; set; }
+
+        public virtual Task Precedence { get; set; } = null!;
+        public virtual Task Task { get; set; } = null!;
+    }
+}
