+++ docker-compose.yml
@@ -12,4 +12,3 @@ services:
     build:
       context: .
       dockerfile: JiraSchedulingConnectAppService/Dockerfile
-
